requirements.txt
Flask
pywinrm
python-dotenv
uvicorn
asgiref
